Update list of required packages to add dependencies encountered building a non standard build host
My attempts to build a Mint based build environment failed for lack of these packages. May not be needed on a supported build host, as they could be part of the default environment for Ubuntu. No Ubuntu host available to me test. They are clearly requirements for a build host though so adding them to the list here does no harm, and could help.